Add fractional display scaling support
Change the rendering pipeline to use float instead of int for the
surface scale factor, enabling fractional scaling (e.g. 150%) on
Wayland and other HiDPI configurations.

- Change RenSurface.scale and RenFont.scale from int to float
- Remove the integer-only assertion in query_surface_scale that
  crashed with fractional scale factors
- Fall back to SDL_GetWindowDisplayScale when pixel/point ratio
  is 1 (Wayland fractional scaling where compositor reports equal
  sizes)
- Create oversized surface when SDL pixel size equals logical size
  to render at native resolution instead of compositor upscaling
- Round font pixel sizes instead of truncating to avoid size
  mismatch at fractional scales (e.g. 22.5 -> 23 not 22)
- Use ceilf for clip rect width/height in scaled_rect to prevent
  sub-pixel clipping at fractional boundaries
- Add explicit int casts throughout the pipeline where float scale
  is multiplied with integer coordinates

Requires building with -Drenderer=true (meson option) which is off
by default on Linux.

+++ src/renderer.h
@@ -20,7 +20,7 @@ typedef enum { FONT_STYLE_BOLD = 1, FONT_STYLE_ITALIC = 2, FONT_STYLE_UNDERLINE 
 typedef struct { uint8_t b, g, r, a; } RenColor;
 typedef struct { int x, y, width, height; } RenRect;
 typedef struct { double offset; } RenTab;
-typedef struct { SDL_Surface *surface; int scale; } RenSurface;
+typedef struct { SDL_Surface *surface; float scale; } RenSurface;
 
 struct RenWindow;
 typedef struct RenWindow RenWindow;
@@ -32,7 +32,7 @@ void ren_font_free(RenFont *font);
 int ren_font_group_get_tab_size(RenFont **font);
 int ren_font_group_get_height(RenFont **font);
 float ren_font_group_get_size(RenFont **font);
-void ren_font_group_set_size(RenFont **font, float size, int surface_scale);
+void ren_font_group_set_size(RenFont **font, float size, float surface_scale);
 #ifdef LITE_USE_SDL_RENDERER
 void update_font_scale(RenWindow *window_renderer, RenFont **fonts);
 #endif
